Discord Human Bot
The Main Purpose of this Bot is to ensure spam robots do not join a Discord Server  purely to Spam. This was made for a Community I am a part of, therefore it is tailored to the needs of that Community and not Necesarily for Public Use. However, I have no problem with having this Code as Open Source.